feat(primitives): leverage `tempo-hardforks` to improve `TempoAddressExt`
