+++ crates/precompiles/src/lib.rs
@@ -74,32 +74,6 @@ pub use tempo_contracts::precompiles::{
 // Re-export storage layout helpers for read-only contexts (e.g., pool validation)
 pub use account_keychain::AuthorizedKey;
 
-/// Fixed system precompile addresses and corresponding activation hardfork
-pub const SYSTEM_PRECOMPILES: &[(Address, TempoHardfork)] = &[
-    (TIP403_REGISTRY_ADDRESS, TempoHardfork::Genesis),
-    (TIP_FEE_MANAGER_ADDRESS, TempoHardfork::Genesis),
-    (STABLECOIN_DEX_ADDRESS, TempoHardfork::Genesis),
-    (NONCE_PRECOMPILE_ADDRESS, TempoHardfork::Genesis),
-    (ACCOUNT_KEYCHAIN_ADDRESS, TempoHardfork::Genesis),
-    (VALIDATOR_CONFIG_ADDRESS, TempoHardfork::Genesis),
-    (VALIDATOR_CONFIG_V2_ADDRESS, TempoHardfork::Genesis),
-    (TIP20_FACTORY_ADDRESS, TempoHardfork::Genesis),
-    (ADDRESS_REGISTRY_ADDRESS, TempoHardfork::T3),
-    (SIGNATURE_VERIFIER_ADDRESS, TempoHardfork::T3),
-    (TIP20_CHANNEL_RESERVE_ADDRESS, TempoHardfork::T5),
-    (RECEIVE_POLICY_GUARD_ADDRESS, TempoHardfork::T6),
-    (STORAGE_CREDITS_ADDRESS, TempoHardfork::T7),
-];
-
-/// Returns `true` if `addr` is any precompile active at `spec`: a TIP-20 token (matched by prefix)
-/// or a fixed system precompile.
-pub fn is_precompile_address(addr: Address, spec: TempoHardfork) -> bool {
-    addr.is_tip20()
-        || SYSTEM_PRECOMPILES
-            .iter()
-            .any(|&(a, activated)| a == addr && spec >= activated)
-}
-
 /// Input per word cost. It covers abi decoding and cloning of input into call data.
 ///
 /// Being careful and pricing it twice as COPY_COST to mitigate different abi decodings.
@@ -1336,25 +1310,6 @@ mod tests {
                 .is_some(),
             "TIP20 channel reserve should be registered at T5"
         );
-    }
-
-    #[test]
-    fn test_is_precompile_address() {
-        for &(address, activated) in SYSTEM_PRECOMPILES {
-            assert!(is_precompile_address(address, activated));
-            assert!(is_precompile_address(address, TempoHardfork::T7));
-
-            if activated != TempoHardfork::Genesis {
-                assert!(!is_precompile_address(address, TempoHardfork::Genesis));
-            }
-        }
-
-        // Assert TIP20 prefixed addresses are classified as precompiles
-        assert!(PATH_USD_ADDRESS.is_tip20());
-        assert!(is_precompile_address(
-            PATH_USD_ADDRESS,
-            TempoHardfork::Genesis
-        ));
     }
 
     #[test]
